requirements.txt
streamlit
tensorflow
opencv-python
pandas
numpy
scikit-learn
pillow